Full song conversion to ViewModel + changing package names  and gradle upgrades due to error with data binding
