+++ .idea/kotlinc.xml
@@ -3,8 +3,4 @@
   <component name="Kotlin2JvmCompilerArguments">
     <option name="jvmTarget" value="1.8" />
   </component>
-  <component name="KotlinCommonCompilerArguments">
-    <option name="apiVersion" value="1.8" />
-    <option name="languageVersion" value="1.8" />
-  </component>
 </project>
