+++ app/src/main/java/com/example/comet/album/AlbumListFromArtistFragment.java
@@ -1,4 +1,4 @@
-package com.example.comet.Song;
+package com.example.comet.album;
 
 import android.content.ContentUris;
 import android.content.Context;
@@ -26,33 +26,31 @@ import android.widget.TextView;
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.request.target.CustomTarget;
 import com.bumptech.glide.request.transition.Transition;
-import com.example.comet.Constants;
+import com.example.comet.song.SongModel;
+import com.example.comet.util.Constants;
 import com.example.comet.R;
-import com.example.comet.UtilMethods;
 
 import java.util.ArrayList;
 import java.util.Locale;
 
-public class SongListFromAlbumFragment extends Fragment {
+public class AlbumListFromArtistFragment extends Fragment implements AlbumRowAdapter.IAlbumRowAdapterInterface {
 
-    private RecyclerView songRecyclerView;
-    private ImageView songAlbumImage;
-    private  ImageView backButton;
-    private TextView songsNumberAndTime;
+    private RecyclerView albumRecyclerView;
+    private ImageView albumArtistImage;
+    private ImageView backButton;
+    private TextView songsNumberAndTimeAlbum;
     private ConstraintLayout midBarLayout;
     private GridLayoutManager gridLayoutManager;
-    private ArrayList<MusicModel> songsList;
+    private ArrayList<AlbumModel> albumsList;
 
-
-    public SongListFromAlbumFragment() {
+    public AlbumListFromArtistFragment() {
         // Required empty public constructor
     }
 
-
-    public static SongListFromAlbumFragment newInstance(ArrayList<MusicModel> songsList) {
-        SongListFromAlbumFragment fragment = new SongListFromAlbumFragment();
+    public static AlbumListFromArtistFragment newInstance(ArrayList<AlbumModel> albumsList) {
+        AlbumListFromArtistFragment fragment = new AlbumListFromArtistFragment();
         Bundle args = new Bundle();
-        args.putParcelableArrayList(Constants.SONGS_PARAM, songsList);
+        args.putParcelableArrayList(Constants.ALBUMS_PARAM, albumsList);
         fragment.setArguments(args);
         return fragment;
     }
@@ -61,7 +59,7 @@ public class SongListFromAlbumFragment extends Fragment {
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         if (getArguments() != null) {
-            songsList = (ArrayList<MusicModel>) getArguments().get(Constants.SONGS_PARAM);
+            albumsList = (ArrayList<AlbumModel>) getArguments().get(Constants.ALBUMS_PARAM);
         }
     }
 
@@ -69,22 +67,20 @@ public class SongListFromAlbumFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        View view = inflater.inflate(R.layout. fragment_song_list_from_album, container, false);
+        View view = inflater.inflate(R.layout.fragment_album_list_from_artist, container, false);
 
-        songAlbumImage = view.findViewById(R.id.artistAlbumImageSong);
-        songsNumberAndTime = view.findViewById(R.id.songsNumberAndTime);
-        midBarLayout = view.findViewById(R.id.midbarLayoutSong);
-        backButton = view.findViewById(R.id.backButtonSong);
-
-
+        albumArtistImage = view.findViewById(R.id.artistAlbumImageAlbum);
+        songsNumberAndTimeAlbum = view.findViewById(R.id.songsNumberAndTimeAlbum);
+        midBarLayout = view.findViewById(R.id.midbarLayoutAlbum);
+        backButton = view.findViewById(R.id.backButtonAlbum);
 
         Uri uri = ContentUris.withAppendedId(Constants.sArtworkUri,
-                Long.parseLong(songsList.get(0).getAlbumId()));
+                Long.parseLong(albumsList.get(0).getId()));
         Glide.with(getContext()).asBitmap().load(uri).placeholder(R.drawable.background_for_load)
                 .error(R.drawable.hoshi).centerCrop().into(new CustomTarget<Bitmap>() {
                     @Override
                     public void onResourceReady(@NonNull Bitmap resource, @Nullable Transition<? super Bitmap> transition) {
-                        songAlbumImage.setImageBitmap(resource);
+                        albumArtistImage.setImageBitmap(resource);
                         Palette.from(resource).generate(new Palette.PaletteAsyncListener() {
                             public void onGenerated(Palette palette) {
                                 Palette.Swatch swatch = palette.getVibrantSwatch();
@@ -92,10 +88,10 @@ public class SongListFromAlbumFragment extends Fragment {
                                 if (swatch != null) {
                                     //swatch.getRgb in both slots will create a solid effect, can find a better color for gradient if wanted
                                     gd = new GradientDrawable(GradientDrawable.Orientation.BOTTOM_TOP, new int[]{swatch.getRgb(), swatch.getRgb()});
-                                    songsNumberAndTime.setTextColor(swatch.getTitleTextColor());
+                                    songsNumberAndTimeAlbum.setTextColor(swatch.getTitleTextColor());
                                 }else {
                                     gd = new GradientDrawable(GradientDrawable.Orientation.BOTTOM_TOP, new int[]{0x38393838, 0x38393838});
-                                    songsNumberAndTime.setTextColor(Color.BLACK);
+                                    songsNumberAndTimeAlbum.setTextColor(Color.BLACK);
                                 }
                                 midBarLayout.setBackground(gd);
                             }
@@ -109,24 +105,26 @@ public class SongListFromAlbumFragment extends Fragment {
                     }
                 });
 
-        long totalSongLength = 0;
-        for(MusicModel song : songsList){
-            totalSongLength += Long.parseLong(song.getDuration());
+        long numAlbumSongs = 0;
+        for(AlbumModel album : albumsList){
+            numAlbumSongs += Long.parseLong(album.getNumberOfSongs());
         }
+        String songsText;
+        String albumsText;
 
         //setting variable text
-        String songsText = (songsList.size() < 2) ? "song" : "songs";
+        albumsText = (albumsList.size() < 2) ? "album" : "albums";
+        songsText = (numAlbumSongs < 2) ? "song" : "songs";
+        songsNumberAndTimeAlbum.setText(String.format(Locale.getDefault(), "%d %s, %d %s", albumsList.size(), albumsText, numAlbumSongs, songsText));
 
-        songsNumberAndTime.setText(String.format(Locale.getDefault(), "%d %s (%s)", songsList.size(), songsText, UtilMethods.prettyDuration(totalSongLength)));
-
-        //using custom adapter to set array of songs into layout
-        songRecyclerView = view.findViewById(R.id.songListFromAlbumRecyclerView);
-        songRecyclerView.setHasFixedSize(true);
+        //using custom adapter to set array of albums into layout
+        albumRecyclerView = view.findViewById(R.id.albumListFromArtistRecyclerView);
+        albumRecyclerView.setHasFixedSize(true);
         gridLayoutManager = new GridLayoutManager(getContext(), 1);
-        songRecyclerView.setAdapter(new SongAdapter(songsList, getContext()));
-        songRecyclerView.setLayoutManager(gridLayoutManager);
+        albumRecyclerView.setAdapter(new AlbumRowAdapter(albumsList, getContext(), this));
+        albumRecyclerView.setLayoutManager(gridLayoutManager);
 
-        backButton.setOnClickListener(v -> moveBackFromAlbum());
+        backButton.setOnClickListener(v -> moveBackfromArtistAlbum());
 
         return view;
     }
@@ -134,26 +132,33 @@ public class SongListFromAlbumFragment extends Fragment {
     @Override
     public void onResume() {
         super.onResume();
-        if(songRecyclerView != null){
-            songRecyclerView.setAdapter(new SongAdapter(songsList, getContext()));
+        if(albumRecyclerView != null){
+            albumRecyclerView.setAdapter(new AlbumRowAdapter(albumsList, getContext(), this));
         }
     }
 
-    public void moveBackFromAlbum(){
+
+    public void moveBackfromArtistAlbum(){
+        mListener.moveBackFromArtistAlbum();
+    }
+    @Override
+    public void toSongsListFromAlbumRow(ArrayList<SongModel> songsList){
         //gets argument from the adapter and calls the method to take back to the main activity
-        mListener.moveBackFromAlbum();
+        mListener.toSongsListFromAlbumRowFragment(songsList);
     }
 
-    SongListFromAlbumFragmentListener mListener;
+    AlbumListFromArtistFragmentListener mListener;
 
     @Override
     public void onAttach(@NonNull Context context) {
         super.onAttach(context);
-        mListener = (SongListFromAlbumFragmentListener) context;
+        mListener = (AlbumListFromArtistFragmentListener) context;
     }
 
-    public interface SongListFromAlbumFragmentListener{
-        //moves back from the songs list inside an album
-        void moveBackFromAlbum();
+    public interface AlbumListFromArtistFragmentListener{
+        //method to send songsList back to the main activity
+        void toSongsListFromAlbumRowFragment(ArrayList<SongModel> songsList);
+        void moveBackFromArtistAlbum();
     }
+
 }
