+++ app/src/main/java/com/example/comet/playlist/PlaylistAdapter.java
@@ -1,4 +1,4 @@
-package com.example.comet.Playlist;
+package com.example.comet.playlist;
 
 import android.content.Context;
 import android.view.LayoutInflater;
