+++ app/src/main/java/com/example/comet/util/Constants.java
@@ -1,4 +1,4 @@
-package com.example.comet;
+package com.example.comet.util;
 
 import android.net.Uri;
