+++ app/src/main/java/com/example/comet/util/InitialData.java
@@ -1,4 +1,4 @@
-package com.example.comet;
+package com.example.comet.util;
 
 import java.io.Serializable;
 import java.util.ArrayList;
