+++ app/src/main/java/com/example/comet/viewmodel/SharedMusicViewModel.java
@@ -1,4 +1,4 @@
-package com.example.comet.ViewModel;
+package com.example.comet.viewmodel;
 
 public class SharedMusicViewModel {
 }
