+++ app/src/main/res/layout/fragment_album.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     android:background="#393838"
-    tools:context=".Song.SongFragment">
+    tools:context=".song.SongFragment">
 
 
     <androidx.recyclerview.widget.RecyclerView
