+++ app/src/main/res/layout/fragment_playlist.xml
@@ -7,7 +7,7 @@
         android:layout_height="match_parent"
         android:fitsSystemWindows="true"
         android:background="#393838"
-        tools:context=".Playlist.PlaylistFragment">
+        tools:context=".playlist.PlaylistFragment">
 
 
         <androidx.recyclerview.widget.RecyclerView
