+++ app/src/main/res/layout/fragment_song_list_from_album.xml
@@ -8,7 +8,7 @@
     android:fitsSystemWindows="true"
     android:clickable="true"
     android:background="@color/main_background"
-    tools:context=".Song.SongListFromAlbumFragment">
+    tools:context=".song.SongListFromAlbumFragment">
 
     <RelativeLayout
         android:id="@+id/relativeLayoutSong"
