+++ app/src/main/res/layout/song_row_item.xml
@@ -1,57 +1,68 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="70dp"
-    app:cardBackgroundColor="@color/main_background">
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+xmlns:app="http://schemas.android.com/apk/res-auto">
 
-    <androidx.constraintlayout.widget.ConstraintLayout
+    <data>
+        <variable
+            name="song"
+            type="com.example.comet.song.SongModel" />
+        <import
+            type="com.example.comet.util.UtilMethods" />
+    </data>
+
+    <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="70dp"
+        app:cardBackgroundColor="@color/main_background">
 
-        <TextView
-            android:id="@+id/songNameText"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="20dp"
-            android:layout_marginTop="13dp"
-            android:text="TextView"
-            android:textColor="@color/white"
-            app:layout_constraintStart_toEndOf="@+id/songImage"
-            app:layout_constraintTop_toTopOf="parent" />
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
 
-        <TextView
-            android:id="@+id/songArtistText"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="20dp"
-            android:layout_marginTop="8dp"
-            android:text="TextView"
-            android:textColor="@color/white"
-            app:layout_constraintStart_toEndOf="@+id/songImage"
-            app:layout_constraintTop_toBottomOf="@+id/songNameText" />
-
-        <ImageView
-            android:id="@+id/songImage"
-            android:layout_width="48dp"
-            android:layout_height="48dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:srcCompat="@drawable/chiaki" />
-
-        <TextView
-            android:id="@+id/songDurationText"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:layout_marginTop="8dp"
-            android:text="TextView"
-            android:textColor="@color/white"
-            app:layout_constraintStart_toEndOf="@+id/songArtistText"
-            app:layout_constraintTop_toBottomOf="@+id/songNameText" />
+            <TextView
+                android:id="@+id/songNameText"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="20dp"
+                android:layout_marginTop="13dp"
+                android:text="@{song.title}"
+                android:textColor="@color/white"
+                app:layout_constraintStart_toEndOf="@+id/songImage"
+                app:layout_constraintTop_toTopOf="parent" />
 
 
-    </androidx.constraintlayout.widget.ConstraintLayout>
+            <TextView
+                android:id="@+id/songArtistText"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="20dp"
+                android:layout_marginTop="8dp"
+                android:text="@{song.artist}"
+                android:textColor="@color/white"
+                app:layout_constraintStart_toEndOf="@+id/songImage"
+                app:layout_constraintTop_toBottomOf="@+id/songNameText" />
 
-</androidx.cardview.widget.CardView>
+            <ImageView
+                android:id="@+id/songImage"
+                android:layout_width="48dp"
+                android:layout_height="48dp"
+                app:albumArt="@{song.albumId}"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="parent" />
+
+            <TextView
+                android:id="@+id/songDurationText"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="8dp"
+                android:layout_marginTop="8dp"
+                android:text="@{UtilMethods.prettyDuration(Long.parseLong(song.duration))}"
+                android:textColor="@color/white"
+                app:layout_constraintStart_toEndOf="@+id/songArtistText"
+                app:layout_constraintTop_toBottomOf="@+id/songNameText" />
+
+
+        </androidx.constraintlayout.widget.ConstraintLayout>
+
+    </androidx.cardview.widget.CardView>
+</layout>
