+++ build.gradle.kts
@@ -5,7 +5,7 @@ buildscript {
         mavenCentral()
     }
     dependencies {
-        classpath("com.android.tools.build:gradle:8.1.4")
+        classpath("com.android.tools.build:gradle:8.7.3")
     }
 }
 
@@ -15,5 +15,5 @@ buildscript {
 //}
 
 plugins {
-    id("com.android.application") version "8.1.4" apply false
+    id("com.android.application") version "8.7.3" apply false
 }
